interface, impl 분리
