+++ src/main/java/com/nexus/sion/feature/squad/command/application/service/SquadCommentCommandService.java
@@ -1,52 +1,10 @@
 package com.nexus.sion.feature.squad.command.application.service;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.nexus.sion.exception.BusinessException;
-import com.nexus.sion.exception.ErrorCode;
-import com.nexus.sion.feature.notification.command.application.service.NotificationCommandService;
 import com.nexus.sion.feature.squad.command.application.dto.request.SquadCommentRegisterRequest;
-import com.nexus.sion.feature.squad.command.domain.aggregate.entity.SquadComment;
-import com.nexus.sion.feature.squad.command.repository.SquadCommentRepository;
 
-import lombok.RequiredArgsConstructor;
+public interface SquadCommentCommandService {
 
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class SquadCommentCommandService {
+  void registerComment(String squadCode, SquadCommentRegisterRequest request, String employeeIdentificationNumber);
 
-  private final SquadCommentRepository squadCommentRepository;
-  private final NotificationCommandService notificationCommandService;
-
-  @Transactional
-  public void registerComment(
-      String squadCode, SquadCommentRegisterRequest request, String employeeIdentificationNumber) {
-    if (request.getContent() == null || request.getContent().trim().isEmpty()) {
-      throw new BusinessException(ErrorCode.COMMENT_CONTENT_EMPTY);
-    }
-
-    SquadComment comment =
-        SquadComment.builder()
-            .squadCode(squadCode)
-            .employeeIdentificationNumber(employeeIdentificationNumber)
-            .content(request.getContent())
-            .build();
-
-    squadCommentRepository.save(comment);
-  }
-
-  public void deleteComment(String squadCode, Long commentId) {
-    SquadComment comment =
-        squadCommentRepository
-            .findById(commentId)
-            .orElseThrow(() -> new BusinessException(ErrorCode.COMMENT_NOT_FOUND));
-
-    if (!comment.getSquadCode().equals(squadCode)) {
-      throw new BusinessException(ErrorCode.INVALID_COMMENT_ACCESS);
-    }
-
-    squadCommentRepository.delete(comment);
-  }
+  void deleteComment(String squadCode, Long commentId);
 }
