+++ src/main/java/com/nexus/sion/feature/squad/command/application/service/SquadCommentCommandServiceImpl.java
@@ -1,0 +1,50 @@
+package com.nexus.sion.feature.squad.command.application.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.nexus.sion.exception.BusinessException;
+import com.nexus.sion.exception.ErrorCode;
+import com.nexus.sion.feature.notification.command.application.service.NotificationCommandService;
+import com.nexus.sion.feature.squad.command.application.dto.request.SquadCommentRegisterRequest;
+import com.nexus.sion.feature.squad.command.domain.aggregate.entity.SquadComment;
+import com.nexus.sion.feature.squad.command.repository.SquadCommentRepository;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class SquadCommentCommandServiceImpl implements SquadCommentCommandService {
+
+  private final SquadCommentRepository squadCommentRepository;
+  private final NotificationCommandService notificationCommandService;
+
+  @Override
+  public void registerComment(
+          String squadCode, SquadCommentRegisterRequest request, String employeeIdentificationNumber) {
+    if (request.getContent() == null || request.getContent().trim().isEmpty()) {
+      throw new BusinessException(ErrorCode.COMMENT_CONTENT_EMPTY);
+    }
+
+    SquadComment comment = SquadComment.builder()
+            .squadCode(squadCode)
+            .employeeIdentificationNumber(employeeIdentificationNumber)
+            .content(request.getContent())
+            .build();
+
+    squadCommentRepository.save(comment);
+  }
+
+  @Override
+  public void deleteComment(String squadCode, Long commentId) {
+    SquadComment comment = squadCommentRepository.findById(commentId)
+            .orElseThrow(() -> new BusinessException(ErrorCode.COMMENT_NOT_FOUND));
+
+    if (!comment.getSquadCode().equals(squadCode)) {
+      throw new BusinessException(ErrorCode.INVALID_COMMENT_ACCESS);
+    }
+
+    squadCommentRepository.delete(comment);
+  }
+}
